+++ src/card/sets/y1995/mod.rs
@@ -1,5 +1,7 @@
 //! Card sets first released in 1995.
 
+pub(super) mod chronicles;
+pub(super) mod fourth_edition;
 pub(super) mod ice_age;
 
 pub(super) use super::{CardRecord, PrintingRecord};