+++ src/card/sets/y1996/mod.rs
@@ -1,5 +1,6 @@
 //! Card sets first released in 1996.
 
+pub(super) mod alliances;
 pub(super) mod mirage;
 
 pub(super) use super::{CardRecord, PrintingRecord};